CI: Move ci/get_hdf5.py to Windows-only
